requirements.txt
alembic
bcrypt==3.1.7
certifi
charset-normalizer
cloudinary
fastapi
Faker
greenlet
httpx
idna
itsdangerous
Jinja2
MarkupSafe
passlib==1.7.4
psycopg2-binary
pydantic
pydantic-core
pydantic-extra-types
pydantic-settings
python-dotenv
python-jose
python-multipart
pytz
PyYAML
requests
six
SQLAlchemy
starlette
urllib3
uvicorn
fastapi-mail
libgravatar
pytest
pytest-cov
redis
fakeredis
sphinx
httpx
fastapi-limiter
pytest-asyncio